CB-1420 Rework CRN parse exceptions
Code for parsing CRN strings, which used to throw either
NullPointerException, IllegalArgumentException, or IllegalStateException
depending on the nature of the parsing problem, is reworked to throw a
new CrnParseException for any parse problems. The code does still
throw NullPointerException, but only if the input string is null for
certain "unsafe" parsing methods.

The CrnTest unit test is moved over to the common module to live near
the class it tests, and also expanded.

UmsAuthenticationService and SdxService are modified slightly to take
advantage of the updated CRN parsing exceptions. This is a behavior
change for both services: they now throw UmsAuthenticationException or
BadRequestException, respectively, for any CRN parsing problems, and
not just for CRNs that do not match the expected CRN pattern.